+++ src/js/profiling/types.ts
@@ -1,6 +1,7 @@
-import type { ThreadCpuProfile } from '@sentry/types';
+import type { ThreadCpuFrame, ThreadCpuProfile } from '@sentry/types';
 
 export interface RawThreadCpuProfile extends ThreadCpuProfile {
+  frames: ThreadCpuFrame[];
   profile_id?: string;
   active_thread_id: string;
 }