+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Automatically detect environment if not set ([#3362](https://github.com/getsentry/sentry-react-native/pull/3362))
 - Send Source Maps Debug ID for symbolicated Profiles ([#3343](https://github.com/getsentry/sentry-react-native/pull/3343))
 
 ### Fixes