+++ CHANGELOG.md
@@ -14,9 +14,9 @@
 
 ### Dependencies
 
-- Bump JavaScript SDK from v7.73.0 to v7.75.1 ([#3344](https://github.com/getsentry/sentry-react-native/pull/3344))
-  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7751)
-  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.73.0...7.75.1)
+- Bump JavaScript SDK from v7.73.0 to v7.76.0 ([#3344](https://github.com/getsentry/sentry-react-native/pull/3344), [#3365](https://github.com/getsentry/sentry-react-native/pull/3365))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7760)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.73.0...7.76.0)
 - Bump Cocoa SDK from v8.13.0 to v8.14.2 ([#3340](https://github.com/getsentry/sentry-react-native/pull/3340))
   - [changelog](https://github.com/getsentry/sentry-cocoa/blob/main/CHANGELOG.md#8142)
   - [diff](https://github.com/getsentry/sentry-cocoa/compare/8.13.0...8.14.2)